+++ tests/testthat/test-rq.R
@@ -6,10 +6,12 @@ if (suppressWarnings(
 )) {
 
   data(stackloss)
-  m1 <- rq(stack.loss ~ Air.Flow + Water.Temp, data = stackloss, tau = .25)
+  m1 <- rq(stack.loss ~ Air.Flow + Water.Temp, data = stackloss, tau = 0.25)
 
   test_that("ggpredict, rq", {
-    expect_warning(pr <- ggpredict(m1, "Air.Flow"))
+    expect_warning({
+      pr <- ggpredict(m1, "Air.Flow")
+    })
     expect_equal(pr$predicted[1], 10.09524, tolerance = 1e-4)
   })