+++ R/vcov.R
@@ -196,6 +196,15 @@ vcov.ggeffects <- function(object, vcov.fun = NULL, vcov.type = NULL, vcov.args 
         insight::check_if_installed("sandwich")
         robust_package <- "sandwich"
       }
+      # clubSandwich does not work for pscl models
+      if (robust_package == "clubSandwich" && inherits(model, c("zeroinfl", "hurdle", "zerotrunc"))) {
+        insight::format_alert(paste0(
+          "Can't compute robust standard errors for models of class `",
+          class(model)[1],
+          "` when `vcov.fun=\"vcovCR\". Please use `vcov.fun=\"vcovCL\"` from the {sandwich} package instead."
+        ))
+        return(NULL)
+      }
       # compute robust standard errors based on vcov
       if (robust_package == "sandwich") {
         vcov.fun <- get(vcov.fun, asNamespace("sandwich"))
@@ -204,6 +213,10 @@ vcov.ggeffects <- function(object, vcov.fun = NULL, vcov.type = NULL, vcov.args 
         vcov.fun <- clubSandwich::vcovCR
         vcm <- as.matrix(do.call(vcov.fun, c(list(obj = model, type = vcov.type), vcov.args)))
       }
+    }
+    # for zero-inflated models, remove zero-inflation part from vcov
+    if (inherits(model, c("zeroinfl", "hurdle", "zerotrunc"))) {
+      vcm <- vcm[!startsWith(rownames(vcm), "zero_"), !startsWith(rownames(vcm), "zero_"), drop = FALSE]
     }
   } else {
     # get variance-covariance-matrix, depending on model type